feat(web): move /icon/{c}/{n}.{svg,mx} to Astro server endpoints
