+++ apps/web/src/pages/icon/[collection]/[name].mx.ts
@@ -1,7 +1,6 @@
-import { hashSha256, isUnsafeSvg, loadCollection } from '../../../src/lib/iconify-cache.ts';
-
-type Env = { ICONS: R2Bucket };
-type Ctx = { request: Request; env: Env; params: { collection: string; name: string } };
+import { svgToMxLibrary } from '@icon-collection/ui';
+import type { APIRoute } from 'astro';
+import { hashSha256, isUnsafeSvg, loadCollection } from '../../../lib/iconify-cache.ts';
 
 // `lib.dom.d.ts`'s `CacheStorage` (pulled in for browser-side Astro/Preact code
 // elsewhere in this app) shadows the Workers runtime's `caches` global, which
@@ -18,12 +17,16 @@ type WorkerCacheStorage = {
 };
 const workerCaches = caches as unknown as WorkerCacheStorage;
 
-export const onRequest = async ({ request, env, params }: Ctx): Promise<Response> => {
+export const prerender = false;
+
+export const GET: APIRoute = async ({ request, locals, params }) => {
+  const env = locals.runtime.env;
   const cache = workerCaches.default;
   const cached = await cache.match(request);
   if (cached) return cached;
 
-  const { collection, name } = params;
+  const collection = params.collection ?? '';
+  const name = params.name ?? '';
   const json = await loadCollection(env, collection);
   if (!json) return new Response('not found', { status: 404 });
   const icon = json.icons[name];
@@ -34,10 +37,11 @@ export const onRequest = async ({ request, env, params }: Ctx): Promise<Response
   const width = icon.width ?? json.width ?? 24;
   const height = icon.height ?? json.height ?? 24;
   const svg = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 ${width} ${height}">${icon.body}</svg>`;
-  const etag = `"sha256:${await hashSha256(svg)}"`;
-  const res = new Response(svg, {
+  const mx = svgToMxLibrary(svg);
+  const etag = `"sha256:${await hashSha256(mx)}"`;
+  const res = new Response(mx, {
     headers: {
-      'content-type': 'image/svg+xml; charset=utf-8',
+      'content-type': 'application/xml; charset=utf-8',
       'cache-control': 'public, max-age=31536000, immutable',
       etag,
     },
