+++ apps/web/src/pages/icon/[collection]/[name].svg.ts
@@ -1,8 +1,5 @@
-import { svgToMxLibrary } from '@icon-collection/ui';
-import { hashSha256, isUnsafeSvg, loadCollection } from '../../../src/lib/iconify-cache.ts';
-
-type Env = { ICONS: R2Bucket };
-type Ctx = { request: Request; env: Env; params: { collection: string; name: string } };
+import type { APIRoute } from 'astro';
+import { hashSha256, isUnsafeSvg, loadCollection } from '../../../lib/iconify-cache.ts';
 
 // `lib.dom.d.ts`'s `CacheStorage` (pulled in for browser-side Astro/Preact code
 // elsewhere in this app) shadows the Workers runtime's `caches` global, which
@@ -19,12 +16,16 @@ type WorkerCacheStorage = {
 };
 const workerCaches = caches as unknown as WorkerCacheStorage;
 
-export const onRequest = async ({ request, env, params }: Ctx): Promise<Response> => {
+export const prerender = false;
+
+export const GET: APIRoute = async ({ request, locals, params }) => {
+  const env = locals.runtime.env;
   const cache = workerCaches.default;
   const cached = await cache.match(request);
   if (cached) return cached;
 
-  const { collection, name } = params;
+  const collection = params.collection ?? '';
+  const name = params.name ?? '';
   const json = await loadCollection(env, collection);
   if (!json) return new Response('not found', { status: 404 });
   const icon = json.icons[name];
@@ -35,11 +36,10 @@ export const onRequest = async ({ request, env, params }: Ctx): Promise<Response
   const width = icon.width ?? json.width ?? 24;
   const height = icon.height ?? json.height ?? 24;
   const svg = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 ${width} ${height}">${icon.body}</svg>`;
-  const mx = svgToMxLibrary(svg);
-  const etag = `"sha256:${await hashSha256(mx)}"`;
-  const res = new Response(mx, {
+  const etag = `"sha256:${await hashSha256(svg)}"`;
+  const res = new Response(svg, {
     headers: {
-      'content-type': 'application/xml; charset=utf-8',
+      'content-type': 'image/svg+xml; charset=utf-8',
       'cache-control': 'public, max-age=31536000, immutable',
       etag,
     },
